requirements.txt
# Core LangChain
langchain==0.2.16
langchain-groq==0.1.9
langchain-pinecone==0.1.3
langchain-community==0.2.16
langchain-huggingface==0.0.3

# FREE Embeddings (local, no API key needed)
sentence-transformers==3.1.1

# Vector DB
pinecone-client>=5.0.0,<6.0.0

# PDF Parsing
pypdf==4.3.1
pdfplumber==0.11.4

# API
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.9
pydantic==2.8.2

# UI
streamlit==1.38.0
requests==2.32.3

# Utils
python-dotenv==1.0.1
tiktoken==0.7.0